+++ src/app/core/@ngrx/app.state.ts
@@ -1,0 +1,5 @@
+import type { ProductsState } from '.';
+
+export interface AppState {
+  products: ProductsState;
+}