+++ src/app/core/@ngrx/products/products.state.ts
@@ -1,0 +1,15 @@
+import { IProduct } from 'src/app/shared/models/product.model';
+
+export interface ProductsState {
+  products: Readonly<IProduct[]>;
+  initialProduct: IProduct | null;
+  readonly loading: boolean;
+  readonly error: Error | string | null;
+}
+
+export const initialProductsState: ProductsState = {
+  products: [],
+  loading: false,
+  error: null,
+  initialProduct: null,
+};