+++ src/app/core/@ngrx/router/router.custom-serializer.ts
@@ -1,0 +1,18 @@
+import { ActivatedRouteSnapshot, RouterStateSnapshot } from '@angular/router';
+import { RouterStateSerializer } from '@ngrx/router-store';
+
+import { RouterStateUrl } from './router.state';
+
+export class CustomSerializer implements RouterStateSerializer<RouterStateUrl> {
+  serialize(routerState: RouterStateSnapshot): RouterStateUrl {
+    let route: ActivatedRouteSnapshot = routerState.root;
+    while (route.firstChild) {
+      route = route.firstChild;
+    }
+
+    const { url } = routerState;
+    const { params } = route;
+
+    return { url, params };
+  }
+}