+++ src/app/core/@ngrx/router/router.selectors.ts
@@ -1,0 +1,6 @@
+import { RouterReducerState } from '@ngrx/router-store';
+import { createFeatureSelector } from '@ngrx/store';
+import { RouterStateUrl } from '.';
+
+export const selectRouterState =
+  createFeatureSelector<RouterReducerState<RouterStateUrl>>('router');